+++ yazi-widgets/src/input/gait.rs
@@ -1,0 +1,10 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Copy, Debug, Default, Deserialize, Eq, PartialEq, Serialize)]
+#[serde(rename_all = "kebab-case")]
+pub enum Gait {
+	#[default]
+	Fine,
+	Lean,
+	Wide,
+}