+++ CHANGELOG.md
@@ -23,6 +23,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/):
 - Rename `<BackTab>` to `<S-Tab>` ([#3989])
 - Remove Legacy Console Mode on Windows ([#3989])
 
+### Deprecated
+
+- Deprecate `backward --far` and `forward --far` in favor of `backward wide` and `forward wide`, respectively ([#4012])
+
 ## [v26.5.6]
 
 ### Added
@@ -1731,3 +1735,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/):
 [#3989]: https://github.com/sxyazi/yazi/pull/3989
 [#3990]: https://github.com/sxyazi/yazi/pull/3990
 [#4005]: https://github.com/sxyazi/yazi/pull/4005
+[#4012]: https://github.com/sxyazi/yazi/pull/4012